The 17th commit.

+++ safe_demo/src/main/java/com/temp/dao/BoxDao.java
@@ -29,7 +29,7 @@ public interface BoxDao {
 	 * @param cardRfid
 	 * @return boolean
 	 */
-	boolean setBoxCardRelationship(int boxId, String cardRfid);
+	boolean setBoxCardRelationship(long boxId, String cardRfid);
 	
 	/**
 	 * setBoxStatusChangeDetails.
